nject: add MatchToEmptyInterface
