+++ nject/types.go
@@ -91,3 +91,5 @@ var terminalErrorType = reflect.TypeOf((*TerminalError)(nil)).Elem()
 var errorType = reflect.TypeOf((*error)(nil)).Elem()
 
 var ignoreType = reflect.TypeOf((*ignore)(nil)).Elem()
+
+var emptyInterfaceType = reflect.TypeOf((*interface{})(nil)).Elem()
